requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
gspread>=5.12.0
google-auth>=2.23.0
python-telegram-bot>=20.0
schedule>=1.2.0
lxml>=4.9.0
python-dotenv>=1.0.0
selenium>=4.15.0